Renamed commons to utils
